pyproject.toml
[project]
name = "semantic-kernel-with-bing"
version = "0.1.0"
description = "An example of calling Grounding with Bing Search in Semantic Kernel"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "semantic-kernel>=1.0.0",
    "azure-search-documents>=11.4.0",
    "azure-identity>=1.15.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "*/site-packages/*"
]


